Add 4UCON and Philmore terminal blocks up to 16 ways.
The 2- and 3-position terminal blocks can interlock to form larger
sizes. So, I've generated footprints up to 16 positions for both the
3.5mm 4UCON terminal blocks and the 5mm Philmore terminal blocks.

In the Philmore terminal blocks, the part number is TB132 for 2
positions, and TB133 for 3 positions. Since the 4- to 16-position
footprints are for multiple terminal blocks interlocked together, I've
simply called them TB13x, since there is not a specific part number
for them.

Corresponds to this pull request:
https://github.com/KiCad/Connectors_Terminal_Blocks.pretty/pull/13