New script for socket strips
Single row SMD left/right sockets for 2.54 and 2.00 mm added. Updated
for KLC3 and new naming convention.
NOTE: most dimensions are from legacy footprints since there are many
variants available. There are references to dimensions and sources that
I have found that could be useful for adding more variants.

Also included are two wrapper classes in a separate module, this for
some of the core primitives. It allows simpler coding in a somewhat
similar style as CadQuery provides for 3D modeling.