Add fp-name, ep size and heel overwrite parameters
For specialized footprints the default naming convention might
not be enough.
For footprints with relatively large ep sizes, there needs to be
a way to limit the ep size and to remove additional heel fillet.
(Doing this automatically would be quite hard so i decided
to simply provide optional parameters for that.)