Fix parameters.py for Conn_PinSocket
The version in the previous commit was copied verbatim from elsewhere,
but did not load the 'datasheet' property. This fixes that, to make the
parameter loading work again.

Also, the num_pins param was calculated as the total number of pins, whereas
the rest of this script treats it as the number of columns instead.